# backend/openapi.yaml
openapi: 3.1.0
info:
  title: ServeFlow API
  version: 1.0.0
  description: Existing ServeFlow REST API. Organization ownership is derived from the authenticated JWT, never from request query parameters.
servers:
  - url: http://localhost:8080/api/v1
security:
  - bearerAuth: []
paths:
...
  /technicians:
    get:
      operationId: listTechnicians
      summary: List technicians
      description: "Defaults to page 1, 20 items, sorted by created_at descending. Sort fields: created_at, updated_at, name, status."
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/search"
        - name: status
          in: query
          schema:
            type: string
            enum: [ACTIVE, INACTIVE]
        - name: sort
          in: query
          schema:
            type: string
            enum: [created_at, updated_at, name, status]
        - $ref: "#/components/parameters/order"
      responses:
        "200":
          $ref: "#/components/responses/TechnicianList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createTechnician
      summary: Create a technician
      requestBody:
        $ref: "#/components/requestBodies/Technician"
      responses:
        "201":
          description: Created technician.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TechnicianResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /technicians/{id}:
    parameters:
      - $ref: "#/components/parameters/id"
    get:
      operationId: getTechnician
      summary: Get a technician
      responses:
        "200":
          description: Technician.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TechnicianResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: updateTechnician
      summary: Update a technician
      requestBody:
        $ref: "#/components/requestBodies/Technician"
      responses:
        "200":
          description: Updated technician.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TechnicianResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      operationId: deactivateTechnician
      summary: Deactivate a technician
      responses:
        "204":
          description: Deactivated.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /invoices:
    get:
      operationId: listInvoices
      summary: List invoices
      description: "Administrators see invoices in their organization; customers see only invoices linked to their authenticated customer profile. Organization and customer scope are derived from the authenticated identity. Defaults to page 1, 20 items, sorted by issue_date descending. Sort fields: issue_date, due_date, total, status, created_at."
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/search"
        - name: status
          in: query
          schema:
            type: string
            enum: [DRAFT, ISSUED, PARTIALLY_PAID, PAID, CANCELLED]
        - name: sort
          in: query
          schema:
            type: string
            enum: [issue_date, due_date, total, status, created_at]
        - $ref: "#/components/parameters/order"
      responses:
        "200":
          $ref: "#/components/responses/InvoiceList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: The account is not an administrator or a linked customer.
    post:
      operationId: createInvoice
      summary: Create an invoice and queue its notification/email event
      requestBody:
        $ref: "#/components/requestBodies/Invoice"
      responses:
        "201":
          description: Created invoice.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
  /invoices/{id}:
    parameters:
      - $ref: "#/components/parameters/id"
    get:
      operationId: getInvoice
      summary: Get invoice details
      description: Administrators can view invoices in their organization. Customers can view only invoices belonging to their linked customer profile.
      responses:
        "200":
          description: Invoice, items, and payments.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceDetails"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "403":
          description: The account is not an administrator or a linked customer.
    put:
      operationId: updateInvoice
      summary: Update an invoice
      requestBody:
        $ref: "#/components/requestBodies/Invoice"
      responses:
        "200":
          description: Updated invoice.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      operationId: cancelInvoice
      summary: Cancel an invoice
      responses:
        "204":
          description: Cancelled.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /invoices/{id}/payments:
    parameters:
      - $ref: "#/components/parameters/id"
    post:
      operationId: recordPayment
      summary: Record a payment and queue its notification/email event
      requestBody:
        $ref: "#/components/requestBodies/Payment"
      responses:
        "201":
          description: Recorded payment and updated invoice.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /notifications:
    get:
      operationId: listNotifications
      summary: List the authenticated user's notifications
      description: Defaults to page 1, 20 items, sorted by created_at descending. Search matches title and message.
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/search"
        - name: is_read
          in: query
          schema:
            type: boolean
      responses:
        "200":
          $ref: "#/components/responses/NotificationList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /notifications/{id}/read:
    parameters:
      - $ref: "#/components/parameters/id"
    put:
      operationId: markNotificationRead
      summary: Mark a notification read
      responses:
        "200":
          description: Updated notification.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    page:
      name: page
      in: query
      description: Page number, from 1 through 1,000,000.
      schema:
        type: integer
        minimum: 1
        maximum: 1000000
        default: 1
    limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    search:
      name: search
      in: query
      schema:
        type: string
        maxLength: 100
    order:
      name: order
      in: query
      schema:
        type: string
        enum: [asc, desc]
  requestBodies:
...
    Technician:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/TechnicianInput"
    Invoice:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/InvoiceInput"
    Payment:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/PaymentInput"
  schemas:
...
    TechnicianInput:
      type: object
      required: [name]
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        email:
          anyOf:
            - type: string
              const: ""
            - type: string
              format: email
              maxLength: 254
        phone:
          type: string
          maxLength: 50
    InvoiceInput:
      type: object
      required: [customer_id, issue_date, due_date, items]
      additionalProperties: false
      properties:
        customer_id:
          type: string
          format: uuid
        appointment_id:
          type: string
          format: uuid
        issue_date:
          type: string
          format: date
        due_date:
          type: string
          format: date
        items:
          type: array
          minItems: 1
          maxItems: 100
          items:
            $ref: "#/components/schemas/InvoiceItemInput"
        tax:
          $ref: "#/components/schemas/MoneyInput"
        notes:
          type: string
          maxLength: 2000
    InvoiceItemInput:
      type: object
      required: [description, quantity, unit_price]
      additionalProperties: false
      properties:
        description:
          type: string
          minLength: 1
          maxLength: 200
        quantity:
          type: integer
          minimum: 1
          maximum: 1000000
        unit_price:
          $ref: "#/components/schemas/MoneyInput"
    PaymentInput:
      type: object
      required: [amount, payment_date, payment_method]
      additionalProperties: false
      properties:
        amount:
          oneOf:
            - type: number
              exclusiveMinimum: 0
              multipleOf: 0.01
            - type: string
              pattern: "^[0-9]+(\\.[0-9]{1,2})?$"
        payment_date:
          type: string
          format: date
        payment_method:
          type: string
          enum: [CASH, CARD, UPI, BANK_TRANSFER, OTHER]
        reference:
          type: string
          maxLength: 120
        notes:
          type: string
          maxLength: 1000
    MoneyInput:
      oneOf:
        - type: number
          minimum: 0
          multipleOf: 0.01
        - type: string
          pattern: "^[0-9]+(\\.[0-9]{1,2})?$"
...
    Customer:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        notes:
          type: string
...
    Technician:
      type: object
      required: [id, organization_id, name, status, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        organization_id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        status:
          type: string
          enum: [ACTIVE, INACTIVE]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    TechnicianResponse:
      type: object
      required: [technician]
      properties:
        technician:
          $ref: "#/components/schemas/Technician"
    Invoice:
      type: object
      required: [id, invoice_number, customer_id, status, issue_date, due_date, subtotal, tax, total, amount_paid, balance_due, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        invoice_number:
          type: string
        customer_id:
          type: string
          format: uuid
        customer_name:
          type: string
        appointment_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [DRAFT, ISSUED, PARTIALLY_PAID, PAID, CANCELLED]
        issue_date:
          type: string
          format: date
        due_date:
          type: string
          format: date
        subtotal:
          type: number
        tax:
          type: number
        total:
          type: number
        amount_paid:
          type: number
        balance_due:
          type: number
        notes:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    InvoiceResponse:
      type: object
      required: [invoice]
      properties:
        invoice:
          $ref: "#/components/schemas/Invoice"
    InvoiceItem:
      type: object
      required: [id, description, quantity, unit_price, amount]
      properties:
        id:
          type: string
          format: uuid
        description:
          type: string
        quantity:
          type: integer
        unit_price:
          type: number
        amount:
          type: number
    Payment:
      type: object
      required: [id, invoice_id, amount, payment_date, payment_method, created_at]
      properties:
        id:
          type: string
          format: uuid
        invoice_id:
          type: string
          format: uuid
        amount:
          type: number
        payment_date:
          type: string
          format: date
        payment_method:
          type: string
          enum: [CASH, CARD, UPI, BANK_TRANSFER, OTHER]
        reference:
          type: string
        notes:
          type: string
        created_at:
          type: string
          format: date-time
    InvoiceDetails:
      type: object
      required: [invoice, customer, items, payments]
      properties:
        invoice:
          $ref: "#/components/schemas/Invoice"
        customer:
          $ref: "#/components/schemas/Customer"
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItem"
        payments:
          type: array
          items:
            $ref: "#/components/schemas/Payment"
    PaymentResponse:
      type: object
      required: [payment, invoice]
      properties:
        payment:
          $ref: "#/components/schemas/Payment"
        invoice:
          $ref: "#/components/schemas/Invoice"
    Notification:
      type: object
      required: [id, type, title, message, is_read, created_at]
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
        title:
          type: string
        message:
          type: string
        is_read:
          type: boolean
        created_at:
          type: string
          format: date-time
    NotificationResponse:
      type: object
      required: [notification]
      properties:
        notification:
          $ref: "#/components/schemas/Notification"
...
    Error:
      type: object
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: object
          required: [code, message]
          additionalProperties: false
          properties:
            code:
              type: string
            message:
              type: string
    Pagination:
      type: object
      required: [page, limit, total, total_pages]
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
...
    TechnicianList:
      type: object
      required: [technicians, pagination]
      properties:
        technicians:
          type: array
          items:
            $ref: "#/components/schemas/Technician"
        pagination:
          $ref: "#/components/schemas/Pagination"
    InvoiceList:
      type: object
      required: [invoices, pagination]
      properties:
        invoices:
          type: array
          items:
            $ref: "#/components/schemas/Invoice"
        pagination:
          $ref: "#/components/schemas/Pagination"
    NotificationList:
      type: object
      required: [notifications, pagination]
      properties:
        notifications:
          type: array
          items:
            $ref: "#/components/schemas/Notification"
        pagination:
          $ref: "#/components/schemas/Pagination"
  responses:
    BadRequest:
      description: Invalid input.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Authentication required or token invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found in the authenticated organization.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Business-rule or uniqueness conflict.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    TechnicianList:
      description: Technician list.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/TechnicianList"
    InvoiceList:
      description: Invoice list.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/InvoiceList"
    NotificationList:
      description: Notification list for the authenticated user.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/NotificationList"